full functionality - without unit test and without spring security

+++ src/main/java/NGSoft/assignment/Shaytaskmanager/service/TaskService.java
@@ -13,6 +13,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.stream.Collectors;
 
 @Service
 @Data
@@ -23,12 +24,15 @@ public class TaskService {
     @Autowired
     private ApplicationContext context;
 
+    @Autowired
+    private MiscService service;
+
     public List<Task> getAllTasks() {
         return taskRepository.findAll();
     }
 
     public List<Task> getTasksByUser(String userName) {
-        return taskRepository.findByAssignee(userName);
+        return taskRepository.findByAssignee(userName).stream().filter(Task::isVisible).collect(Collectors.toList());
     }
 
     public Task saveTask(TaskWebRequest task) {
@@ -39,25 +43,43 @@ public class TaskService {
         if (!isAssignorAdmin){
             throw new RuntimeException(ExceptionMessages.TASK_OPERATION_ADD_NOT_ALLOWED);
         }
-        return taskRepository.save(MiscService.buildTaskDBObject(task));
+
+        return taskRepository.save(service.buildTaskDBObject(task));
     }
 
-    public Task updateExistingTask(int id, TaskWebRequest task){
+    /**
+     * update general task data: description and title
+     * @param id - task id
+     * @param task
+     * @return
+     */
+    public Task updateGeneralTaskData(int id, TaskWebRequest task){
 
         Task existingTask;
         try {
             existingTask = taskRepository.findById(id).orElseThrow();
         } catch (NoSuchElementException e) {
-            throw new RuntimeException(ExceptionMessages.TASK_NOT_EXIST);
+            throw new RuntimeException(ExceptionMessages.OBJECT_NOT_EXIST);
         }
-        existingTask.setAssignee(task.getAssignee() != null ? task.getAssignee() : existingTask.getAssignee());
+        boolean isAssignorAdmin = context.getBean(UserRepository.class).findByName(task.getAssignor()).getIsAdmin();
+        if (!isAssignorAdmin){
+            throw new RuntimeException(ExceptionMessages.TASK_OPERATION_ADD_NOT_ALLOWED);
+        }
         existingTask.setDescription(task.getDescription() != null ? task.getDescription() : existingTask.getDescription());
         existingTask.setTitle(task.getTitle() != null ? task.getTitle() : existingTask.getTitle());
-        existingTask.setStatus(Status.getStatusById(task.getStatus()));
         return taskRepository.save(existingTask);
     }
 
-    public Task changeTaskStatus(int id, StatusChaneRequest statusChangeRequest) {
+    /**
+     * change task status. if the task already in COMPLETE state, it can only move to ARCHIVED, and only user
+     * with Admin permission can perform this operation
+     * @param id task id
+     * @param statusChangeRequest
+     * @param requester - the user who perform the operation
+     * @return
+     */
+    public Task changeTaskStatus(int id, StatusChaneRequest statusChangeRequest, String requester) {
+        User requesterUser  = context.getBean(UserRepository.class).findByName(requester);
         Task existingTask = taskRepository.findById(id).orElseThrow();
         if (statusChangeRequest.getNewStatus() == existingTask.getStatus().getStatus()) {
             throw new RuntimeException(ExceptionMessages.TASK_NEW_STATUS_SAME_OLD_STATUS);
@@ -67,14 +89,45 @@ public class TaskService {
             if (!(Status.getStatusById(statusChangeRequest.getNewStatus()).equals(Status.ARCHIVED))) {
                 throw new RuntimeException(ExceptionMessages.TASK_INVALID_STATUS_FLOW);
             }
+            else {
+                service.isUserPermittedForTaskOperation(requesterUser, existingTask);
+                existingTask.setStatus(Status.ARCHIVED);
+                existingTask.setVisible(false);
+            }
         }
         existingTask.setStatus(Status.getStatusById(statusChangeRequest.getNewStatus()));
         return taskRepository.save(existingTask);
     }
 
+    /**
+     * delete task, only if user has admin permission.
+     * @param id task id
+     * @return
+     */
     public Task deleteTask(int id) {
         Task existingTask = taskRepository.findById(id).orElseThrow();
+        boolean isAssignorAdmin = context.getBean(UserRepository.class).findByName(existingTask.getAssignor()).getIsAdmin();
+        if (!isAssignorAdmin){
+            throw new RuntimeException(ExceptionMessages.TASK_OPERATION_ADD_NOT_ALLOWED);
+        }
         taskRepository.delete(existingTask);
         return existingTask;
+    }
+
+    /**
+     * update task assignee
+     * @param id task id
+     * @param taskToUpdate
+     * @param requester  the user who perform the operation
+     * @return
+     */
+    public Task updateTaskAssignee(int id, TaskWebRequest taskToUpdate, String requester) {
+        User requesterUser  = context.getBean(UserRepository.class).findByName(requester);
+        Task existingTask = taskRepository.findById(id).orElseThrow();
+
+        service.isUserPermittedForTaskOperation(requesterUser, existingTask);
+
+        existingTask.setAssignee(taskToUpdate.getAssignee());
+        return taskRepository.save(existingTask);
     }
 }
